improve autocomplete and add option to import without globals
